add Visualize Util and migrate visualize structure to there
