+++ scala-package/core/src/main/scala/org/apache/mxnet/NDArray.scala
@@ -201,15 +201,6 @@ object NDArray extends NDArrayBase {
   }
 
   /**
-    * Get the String representation of NDArray
-    * @param nd input NDArray
-    * @return String
-    */
-  def toString(nd : NDArray) : String = {
-    nd.visualize
-  }
-
-  /**
    * Create an empty uninitialized new NDArray, with specified shape.
    *
    * @param shape shape of the NDArray.
@@ -823,53 +814,7 @@ class NDArray private[mxnet](private[mxnet] val handle: NDArrayHandle,
     checkCall(_LIB.mxFloat64NDArraySyncCopyFromCPU(handle, source, source.length))
 
   override def toString() : String = {
-    s"${this.visualize}<NDArray ${this.shape} ${this.context}>"
-  }
-
-  /**
-    * Visualize the internal structure of NDArray
-    * @return String that show the structure
-    */
-  def visualize: String = {
-    buildStringHelper(this, this.shape.length) + "\n"
-  }
-  /**
-    * Helper function to create formatted NDArray output
-    * The NDArray will be represented in a reduced version if too large
-    * @param nd NDArray as the input
-    * @param totalSpace totalSpace of the lowest dimension
-    * @return String format of NDArray
-    */
-  private def buildStringHelper(nd : NDArray, totalSpace : Int) : String = {
-    var result = ""
-    val THRESHOLD = 100000      // longest NDArray to show in full
-    val ARRAYTHRESHOLD = 1000   // longest array to show in full
-    val shape = nd.shape
-    val space = totalSpace - shape.length
-    if (shape.length != 1) {
-      val (length, postfix) =
-        if (shape.product > THRESHOLD) {
-          // reduced NDArray
-          (1, s"\n${" " * (space + 1)}... with length ${shape(0)}\n")
-        } else {
-          (shape(0), "")
-        }
-      for (num <- 0 until length) {
-        val output = buildStringHelper(nd.at(num), totalSpace)
-        result += s"$output\n"
-      }
-      result = s"${" " * space}[\n$result${" " * space}$postfix]"
-    } else {
-      if (shape(0) > ARRAYTHRESHOLD) {
-        // reduced Array
-        val front = nd.slice(0, 10)
-        val back = nd.slice(shape(0) - 10, shape(0) - 1)
-        result = s"${" " * space}[${front.toArray.mkString(",")} ... ${back.toArray.mkString(",")}]"
-      } else {
-        result = s"${" " * space}[${nd.toArray.mkString(",")}]"
-      }
-    }
-    result
+    s"<NDArray ${this.shape} ${this.context}>"
   }
 
   /**
